feat: Choose guess based on the fewest remaining solutions
This assumes that the best guess is the one that has the fewest
remaining options over the entire set of solutions. It calculates
this by grouping the clues returned by applying the word to every
possible solution and tracking how many times each clue is returned.

It then sums the squares of each group count to determine the total
number of options. This is a proxy for the average number of
possible solutions for the word.

The best word is the word that has the lowest sum of options for
all possible solutions.